+++ src/primitives/text.css
@@ -1,4 +1,4 @@
-@import url(./spacing.css);
+@import url(../foundations/spacing.css);
 
 .typography-defaults {
   box-sizing: border-box;
