+++ src/recipes/card.css
@@ -1,4 +1,4 @@
-@import url(./spacing.css);
+@import url(../foundations/spacing.css);
 
 .card {
   display: flex;
