+++ src/stories/Card.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import { Card } from './Card';
+import { Card } from '../recipes/Card';
 
 export default {
   title: 'Example/Card',
